+++ tests/bas_metadata_library/test_standard_iso_19115_1.py
@@ -25,6 +25,7 @@ from bas_metadata_library import RecordValidationError
 from bas_metadata_library.standards.iso_19115_1 import (
     Namespaces,
     MetadataRecordConfigV2,
+    MetadataRecordConfigV3,
     MetadataRecord,
 )
 from bas_metadata_library.standards.iso_19115_common.utils import format_numbers_consistently, encode_date_string
@@ -36,10 +37,7 @@ from tests.bas_metadata_library.standard_iso_19115_1_common import (
     assert_online_resource,
     assert_identifier,
 )
-from tests.resources.configs.iso19115_1_standard import (
-    configs_safe_v2,
-    configs_v2_all,
-)
+from tests.resources.configs.iso19115_1_standard import configs_safe_v2, configs_v2_all, configs_safe_v3
 
 
 MonkeyPatch.patch_fromisoformat()
@@ -52,6 +50,14 @@ def test_invalid_configuration_v2():
     config = {"invalid-configuration": "invalid-configuration"}
     with pytest.raises(ValidationError) as e:
         configuration = MetadataRecordConfigV2(**config)
+        configuration.validate()
+    assert "'hierarchy_level' is a required property" in str(e.value)
+
+
+def test_invalid_configuration_v3():
+    config = {"invalid-configuration": "invalid-configuration"}
+    with pytest.raises(ValidationError) as e:
+        configuration = MetadataRecordConfigV3(**config)
         configuration.validate()
     assert "'hierarchy_level' is a required property" in str(e.value)
 
@@ -116,8 +122,68 @@ def test_configuration_v2_json_round_trip(config_name):
     assert configuration.config == _config.config
 
 
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_configuration_v3_from_json_file(config_name):
+    configuration = MetadataRecordConfigV3()
+    config_path = Path().resolve().parent.joinpath(f"resources/configs/{standard}/{config_name}.json")
+    configuration.load(file=config_path)
+    configuration.validate()
+    assert configuration.config == configs_safe_v3[config_name]
+
+
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_configuration_v3_from_json_string(config_name):
+    with open(Path().resolve().parent.joinpath(f"resources/configs/{standard}/{config_name}.json"), mode="r") as file:
+        config_str = file.read()
+        configuration = MetadataRecordConfigV3()
+        configuration.loads(string=config_str)
+        configuration.validate()
+        assert configuration.config == configs_safe_v3[config_name]
+
+
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_configuration_v3_to_json_file(config_name):
+    configuration = MetadataRecordConfigV3(**configs_safe_v3[config_name])
+
+    with TemporaryDirectory() as tmp_dir_name:
+        config_path = Path(tmp_dir_name).joinpath("config.json")
+        configuration.dump(file=config_path)
+        with open(config_path, mode="r") as config_file:
+            config = json.load(config_file)
+
+        with open(
+            Path().resolve().parent.joinpath(f"resources/configs/{standard}/{config_name}.json"), mode="r"
+        ) as _config_file:
+            _config = json.load(_config_file)
+
+        assert config == _config
+
+
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_configuration_v3_to_json_string(config_name):
+    configuration = MetadataRecordConfigV3(**configs_safe_v3[config_name])
+
+    config = configuration.dumps()
+
+    with open(
+        Path().resolve().parent.joinpath(f"resources/configs/{standard}/{config_name}.json"), mode="r"
+    ) as _config_file:
+        _config = _config_file.read()
+
+    assert config == _config
+
+
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_configuration_v3_json_round_trip(config_name):
+    configuration = MetadataRecordConfigV3(**configs_safe_v3[config_name])
+    config = configuration.dumps()
+    _config = MetadataRecordConfigV3()
+    _config.loads(config)
+    assert configuration.config == _config.config
+
+
 @pytest.mark.usefixtures("app_client")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_response(client, config_name):
     response = client.get(f"/standards/{standard}/{config_name}")
     assert response.status_code == HTTPStatus.OK
@@ -125,7 +191,7 @@ def test_response(client, config_name):
 
 
 @pytest.mark.usefixtures("app_client")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_complete_record(client, config_name):
     with open(
         Path().resolve().parent.joinpath(f"resources/records/{standard}/{config_name}-record.xml"), mode="r"
@@ -136,7 +202,7 @@ def test_complete_record(client, config_name):
 
 
 @pytest.mark.usefixtures("app_client")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_xml_declaration(client, config_name):
     response = client.get(f"/standards/{standard}/{config_name}")
     record = ElementTree(XML(response.data))
@@ -145,7 +211,7 @@ def test_xml_declaration(client, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_xml_namespaces(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
     expected_namespaces = Namespaces().nsmap()
@@ -153,7 +219,7 @@ def test_xml_namespaces(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_root_element(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
 
@@ -162,10 +228,10 @@ def test_root_element(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_file_identifier(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "file_identifier" not in config:
         pytest.skip("record does not contain a file identifier")
@@ -178,10 +244,10 @@ def test_file_identifier(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_language(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "metadata" not in config or "language" not in config["metadata"]:
         pytest.skip("record does not contain a metadata language")
@@ -197,10 +263,10 @@ def test_language(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_character_set(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "metadata" not in config or "character_set" not in config["metadata"]:
         pytest.skip("record does not contain a metadata character set")
@@ -217,10 +283,10 @@ def test_character_set(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_hierarchy_level(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "hierarchy_level" not in config:
         pytest.skip("record does not contain a hierarchy level")
@@ -235,10 +301,10 @@ def test_hierarchy_level(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_hierarchy_level_name(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "hierarchy_level" not in config:
         pytest.skip("record does not contain a hierarchy level name")
@@ -251,10 +317,10 @@ def test_hierarchy_level_name(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_contact(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "metadata" not in config or "contacts" not in config["metadata"]:
         pytest.skip("record does not contain any metadata contacts")
@@ -274,10 +340,10 @@ def test_contact(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_datestamp(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "metadata" not in config or "date_stamp" not in config["metadata"]:
         pytest.skip("record does not contain a metadata datestamp")
@@ -288,10 +354,10 @@ def test_datestamp(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_metadata_standard(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if (
         "metadata" not in config
@@ -309,10 +375,10 @@ def test_metadata_standard(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_metadata_standard_version(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if (
         "metadata" not in config
@@ -330,10 +396,10 @@ def test_metadata_standard_version(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_reference_system_info(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "reference_system_info" not in config:
         pytest.skip("record does not contain coordinate reference system information")
@@ -387,10 +453,10 @@ def test_reference_system_info(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_citation(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config:
         pytest.skip("record does not contain an identification citation")
@@ -404,10 +470,10 @@ def test_identification_citation(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_abstract(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "abstract" not in config["identification"]:
         pytest.skip("record does not contain an identification abstract")
@@ -421,10 +487,10 @@ def test_identification_abstract(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_purpose(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "purpose" not in config["identification"]:
         pytest.skip("record does not contain an identification purpose")
@@ -438,10 +504,10 @@ def test_identification_purpose(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_credit(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "credit" not in config["identification"]:
         pytest.skip("record does not contain an identification credit")
@@ -455,10 +521,10 @@ def test_identification_credit(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_status(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "status" not in config["identification"]:
         pytest.skip("record does not contain an identification status")
@@ -515,10 +581,10 @@ def _resolve_points_of_contact_xpaths(point_of_contact_type, config):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_points_of_contact(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "contacts" not in config["identification"]:
         pytest.skip("record does not contain any identification points of contact")
@@ -543,10 +609,10 @@ def test_identification_points_of_contact(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_maintenance(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "maintenance" not in config["identification"]:
         pytest.skip("record does not contain identification maintenance")
@@ -561,10 +627,10 @@ def test_identification_maintenance(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_graphic_overviews(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "graphic_overviews" not in config["identification"]:
         pytest.skip("record does not contain graphic overviews")
@@ -677,10 +743,10 @@ def _resolve_descriptive_keywords_xpaths(config) -> List[dict]:
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_descriptive_keywords(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "keywords" not in config["identification"]:
         pytest.skip("record does not contain identification keywords")
@@ -819,10 +885,10 @@ def test_identification_aggregations(client, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_spatial_representation_type(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "spatial_representation_type" not in config["identification"]:
         pytest.skip("record does not contain an identification spatial representation type")
@@ -860,18 +926,18 @@ def _test_identification_spatial_resolution(record, config):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_spatial_resolution(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
     _test_identification_spatial_resolution(record=record, config=config)
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_character_set(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "character_set" not in config["identification"]:
         pytest.skip("record does not contain an identification character set")
@@ -885,10 +951,10 @@ def test_identification_character_set(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_language(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "language" not in config["identification"]:
         pytest.skip("record does not contain an identification language")
@@ -904,10 +970,10 @@ def test_identification_language(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_topics(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "topics" not in config["identification"]:
         pytest.skip("record does not contain any ISO topics")
@@ -922,10 +988,10 @@ def test_identification_topics(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_geographic_extent_bounding_box(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if (
         "identification" not in config
@@ -969,10 +1035,10 @@ def test_identification_geographic_extent_bounding_box(get_record_response, conf
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_geographic_extent_identifier(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if (
         "identification" not in config
@@ -995,10 +1061,10 @@ def test_identification_geographic_extent_identifier(get_record_response, config
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_temporal_extent(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if (
         "identification" not in config
@@ -1040,10 +1106,10 @@ def test_identification_temporal_extent(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_vertical_extent(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if (
         "identification" not in config
@@ -1085,10 +1151,10 @@ def test_identification_vertical_extent(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_identification_supplemental_info(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "supplemental_information" not in config["identification"]:
         pytest.skip("record does not contain supplemental information")
@@ -1130,10 +1196,10 @@ def _check_distributors_are_unique(distributions: list, record: MetadataRecord) 
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_distributions(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
     xpath_base = "/gmd:MD_Metadata/gmd:distributionInfo/gmd:MD_Distribution/gmd:distributor/gmd:MD_Distributor"
 
     if "distribution" not in config:
@@ -1210,10 +1276,10 @@ def test_distributions(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_data_quality_scope(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "hierarchy_level" not in config:
         pytest.skip("record does not contain a hierarchy level / scope code")
@@ -1228,10 +1294,10 @@ def test_data_quality_scope(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_data_quality_lineage(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "identification" not in config or "lineage" not in config["identification"]:
         pytest.skip("record does not contain a lineage")
@@ -1245,10 +1311,10 @@ def test_data_quality_lineage(get_record_response, config_name):
 
 
 @pytest.mark.usefixtures("get_record_response")
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_metadata_maintenance(get_record_response, config_name):
     record = get_record_response(standard=standard, config=config_name)
-    config = configs_safe_v2[config_name]
+    config = configs_safe_v3[config_name]
 
     if "metadata" not in config or "maintenance" not in config["metadata"]:
         pytest.skip("record does not contain metadata maintenance")
@@ -1261,12 +1327,12 @@ def test_metadata_maintenance(get_record_response, config_name):
 
 
 def test_edge_case_contact_without_email_address():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["metadata"]["contacts"][0]["address"] = {}
     config["metadata"]["contacts"][0]["address"][
         "delivery_point"
     ] = "British Antarctic Survey, High Cross, Madingley Road"
-    configuration = MetadataRecordConfigV2(**config)
+    configuration = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration)
     document = fromstring(record.generate_xml_document())
     contact_email_value = document.xpath(
@@ -1278,12 +1344,12 @@ def test_edge_case_contact_without_email_address():
 
 
 def test_edge_case_citation_with_multiple_roles():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["reference_system_info"] = {
         "code": {"value": "urn:ogc:def:crs:EPSG::4326"},
         "authority": {"contact": {"individual": {"name": "foo"}, "role": ["publisher", "author"]}},
     }
-    configuration = MetadataRecordConfigV2(**config)
+    configuration = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration)
     with pytest.raises(ValueError) as e:
         record.generate_xml_document()
@@ -1292,11 +1358,11 @@ def test_edge_case_citation_with_multiple_roles():
 
 
 def test_edge_case_identifier_without_href():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["identification"]["identifiers"] = [
         {"identifier": "NE/E007895/1", "namespace": "award"},
     ]
-    configuration = MetadataRecordConfigV2(**config)
+    configuration = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration)
     document = fromstring(record.generate_xml_document())
     identifier_value = document.xpath(
@@ -1310,7 +1376,7 @@ def test_edge_case_identifier_without_href():
 
 def test_edge_case_datestamp_invalid_date():
     with open(
-        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v2-record.xml"), mode="r"
+        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v3-record.xml"), mode="r"
     ) as record_file:
         record_data = record_file.read()
     record_element = XML(record_data.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1326,9 +1392,9 @@ def test_edge_case_datestamp_invalid_date():
 
 
 def test_edge_case_spatial_resolution_null():
-    config = deepcopy(configs_safe_v2["complete_v2"])
+    config = deepcopy(configs_safe_v3["complete_v3"])
     config["identification"]["spatial_resolution"] = None
-    config_ = MetadataRecordConfigV2(**config)
+    config_ = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config_).generate_xml_document()
     record = fromstring(record)
 
@@ -1337,7 +1403,7 @@ def test_edge_case_spatial_resolution_null():
 
 def test_edge_case_date_invalid_date():
     with open(
-        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v2-record.xml"), mode="r"
+        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v3-record.xml"), mode="r"
     ) as record_file:
         record_data = record_file.read()
     record_element = XML(record_data.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1354,7 +1420,7 @@ def test_edge_case_date_invalid_date():
 
 def test_edge_case_temporal_extent_begin_invalid_date():
     with open(
-        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v2-record.xml"), mode="r"
+        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v3-record.xml"), mode="r"
     ) as record_file:
         record_data = record_file.read()
     record_element = XML(record_data.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1371,7 +1437,7 @@ def test_edge_case_temporal_extent_begin_invalid_date():
 
 def test_edge_case_temporal_extent_end_invalid_date():
     with open(
-        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v2-record.xml"), mode="r"
+        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v3-record.xml"), mode="r"
     ) as record_file:
         record_data = record_file.read()
     record_element = XML(record_data.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1388,7 +1454,7 @@ def test_edge_case_temporal_extent_end_invalid_date():
 
 def test_edge_case_temporal_extent_begin_missing_date():
     with open(
-        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v2-record.xml"), mode="r"
+        Path().resolve().parent.joinpath(f"resources/records/{standard}/minimal_v3-record.xml"), mode="r"
     ) as record_file:
         record_data = record_file.read()
     record_element = XML(record_data.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1403,7 +1469,7 @@ def test_edge_case_temporal_extent_begin_missing_date():
 
 
 def test_edge_case_distribution_option_format_no_properties():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["distribution"] = [
         {
             "distributor": {
@@ -1434,7 +1500,7 @@ def test_edge_case_distribution_option_format_no_properties():
             ],
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config).generate_xml_document().decode()
     record_element = XML(record.encode(), parser=XMLParser(remove_blank_text=True))
     record = tostring(record_element).decode()
@@ -1459,7 +1525,7 @@ def test_edge_case_distribution_option_format_no_properties():
 
 
 def test_edge_case_distribution_option_transfer_options_no_properties():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["distribution"] = [
         {
             "distributor": {
@@ -1490,7 +1556,7 @@ def test_edge_case_distribution_option_transfer_options_no_properties():
             ],
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config).generate_xml_document().decode()
     record_element = XML(record.encode(), parser=XMLParser(remove_blank_text=True))
     record = tostring(record_element).decode()
@@ -1504,7 +1570,7 @@ def test_edge_case_distribution_option_transfer_options_no_properties():
 
 
 def test_edge_case_distribution_option_no_id():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["distribution"] = [
         {
             "distributor": {
@@ -1535,7 +1601,7 @@ def test_edge_case_distribution_option_no_id():
             ],
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config).generate_xml_document().decode()
     record_element = XML(record.encode(), parser=XMLParser(remove_blank_text=True))
     record = tostring(record_element).decode()
@@ -1562,7 +1628,7 @@ def test_edge_case_distribution_option_no_id():
 
 
 def test_edge_case_distribution_option_more_formats_than_transfer_options():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["distribution"] = [
         {
             "distributor": {
@@ -1590,7 +1656,7 @@ def test_edge_case_distribution_option_more_formats_than_transfer_options():
             ],
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config)
     del record.attributes["distribution"][0]["distribution_options"][1]["transfer_option"]
     record = record.generate_xml_document().decode()
@@ -1608,7 +1674,7 @@ def test_edge_case_distribution_option_more_formats_than_transfer_options():
 
 
 def test_edge_case_distribution_option_more_transfer_options_than_formats():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["distribution"] = [
         {
             "distributor": {
@@ -1636,7 +1702,7 @@ def test_edge_case_distribution_option_more_transfer_options_than_formats():
             ],
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config)
     del record.attributes["distribution"][0]["distribution_options"][1]["format"]
     record = record.generate_xml_document().decode()
@@ -1652,7 +1718,7 @@ def test_edge_case_distribution_option_more_transfer_options_than_formats():
 
 
 def test_edge_case_distribution_option_transfer_option_size_no_unit():
-    config = deepcopy(configs_safe_v2["minimal_v2"])
+    config = deepcopy(configs_safe_v3["minimal_v3"])
     config["distribution"] = [
         {
             "distributor": {
@@ -1684,7 +1750,7 @@ def test_edge_case_distribution_option_transfer_option_size_no_unit():
             ],
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config)
     record = record.generate_xml_document().decode()
     _record = MetadataRecord(record=record)
@@ -1693,7 +1759,7 @@ def test_edge_case_distribution_option_transfer_option_size_no_unit():
 
 
 def test_edge_case_citation_title_anchor_no_value_with_href():
-    config = deepcopy(configs_safe_v2["complete_v2"])
+    config = deepcopy(configs_safe_v3["complete_v3"])
     config["identification"]["keywords"] = [
         {
             "terms": [
@@ -1722,7 +1788,7 @@ def test_edge_case_citation_title_anchor_no_value_with_href():
             },
         }
     ]
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config)
     record = record.generate_xml_document().decode()
     record_element = XML(record.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1737,9 +1803,9 @@ def test_edge_case_citation_title_anchor_no_value_with_href():
 
 @pytest.mark.parametrize("contact_type", ["individual", "organisation"])
 def test_edge_case_responsible_party_anchor_no_value_with_href(contact_type):
-    config = deepcopy(configs_safe_v2["complete_v2"])
+    config = deepcopy(configs_safe_v3["complete_v3"])
     config["metadata"]["contacts"][0][contact_type] = {"name": "*Name to be removed*", "href": "*Test value*"}
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config)
     record = record.generate_xml_document().decode()
     record_element = XML(record.encode(), parser=XMLParser(remove_blank_text=True))
@@ -1774,9 +1840,9 @@ def test_edge_case_responsible_party_anchor_no_value_with_href(contact_type):
     ],
 )
 def test_edge_case_responsible_party_incomplete_address(address_config):
-    config = deepcopy(configs_safe_v2["complete_v2"])
+    config = deepcopy(configs_safe_v3["complete_v3"])
     config["metadata"]["contacts"][0]["address"] = address_config
-    config = MetadataRecordConfigV2(**config)
+    config = MetadataRecordConfigV3(**config)
     record = MetadataRecord(configuration=config)
     record = record.generate_xml_document().decode()
     _record = MetadataRecord(record=record)
@@ -1785,11 +1851,11 @@ def test_edge_case_responsible_party_incomplete_address(address_config):
 
 
 def test_edge_case_parse_config_year_only_date_no_precision():
-    _config = deepcopy(configs_safe_v2["minimal_v2"])
+    _config = deepcopy(configs_safe_v3["minimal_v3"])
     _config["metadata"]["date_stamp"] = "2018-10-18"
     _config["identification"]["dates"]["creation"] = {"date": "2018"}
     _config = json.dumps(_config)
-    config = MetadataRecordConfigV2()
+    config = MetadataRecordConfigV3()
     config.loads(string=_config)
     assert config.config["identification"]["dates"]["creation"] == {
         "date": date(2018, 1, 1),
@@ -1837,24 +1903,59 @@ def test_lossless_conversion_v2(get_record_response, config_name):
     _config = configs_safe_v2[config_name]
 
     record = MetadataRecord(record=_record)
-    config_ = record.make_config().config
+    config_v3 = record.make_config()
+    config_ = config_v3.downgrade_to_v2_config().config
 
-    config = MetadataRecordConfigV2(**config_)
+    config = MetadataRecordConfigV3()
+    config.upgrade_from_v2_config(v2_config=MetadataRecordConfigV2(**config_))
     record_ = MetadataRecord(configuration=config).generate_xml_document().decode()
     assert _record == record_
     assert _config == config_
 
 
-@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_parse_existing_record_v3(config_name):
+    with open(
+        Path().resolve().parent.joinpath(f"resources/records/{standard}/{config_name}-record.xml"), mode="r"
+    ) as record_file:
+        record_data = record_file.read()
+
+    record = MetadataRecord(record=record_data)
+    configuration = record.make_config()
+    config = configuration.config
+    assert config == configs_safe_v3[config_name]
+
+
+@pytest.mark.usefixtures("get_record_response")
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
+def test_lossless_conversion_v3(get_record_response, config_name):
+    _record = tostring(
+        get_record_response(standard=standard, config=config_name),
+        pretty_print=True,
+        xml_declaration=True,
+        encoding="utf-8",
+    ).decode()
+    _config = configs_safe_v3[config_name]
+
+    record = MetadataRecord(record=_record)
+    config_ = record.make_config().config
+
+    config = MetadataRecordConfigV3(**config_)
+    record_ = MetadataRecord(configuration=config).generate_xml_document().decode()
+    assert _record == record_
+    assert _config == config_
+
+
+@pytest.mark.parametrize("config_name", list(configs_safe_v3.keys()))
 def test_record_schema_validation_valid(config_name):
-    config = MetadataRecordConfigV2(**configs_safe_v2[config_name])
+    config = MetadataRecordConfigV3(**configs_safe_v3[config_name])
     record = MetadataRecord(configuration=config)
     record.validate()
     assert True is True
 
 
 def test_record_schema_validation_invalid():
-    config = deepcopy(MetadataRecordConfigV2(**configs_safe_v2["minimal_v2"]))
+    config = deepcopy(MetadataRecordConfigV3(**configs_safe_v3["minimal_v3"]))
     record = MetadataRecord(configuration=config)
     with pytest.raises(RecordValidationError) as e:
         record.attributes["identification"]["spatial_resolution"] = "invalid"